requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.27.0
anthropic>=0.39.0
httpx>=0.27.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
pyyaml>=6.0
aiohttp>=3.9.0
